+++ tests/data/models/TestAssetParent.php
@@ -6,8 +6,9 @@ use davidhirtz\yii2\media\models\interfaces\AssetParentInterface;
 use davidhirtz\yii2\media\models\traits\AssetParentTrait;
 use davidhirtz\yii2\skeleton\db\ActiveQuery;
 use davidhirtz\yii2\skeleton\db\ActiveRecord;
+use davidhirtz\yii2\skeleton\models\interfaces\TypeAttributeInterface;
 
-class TestAssetParent extends ActiveRecord implements AssetParentInterface
+class TestAssetParent extends ActiveRecord implements AssetParentInterface, TypeAttributeInterface
 {
     use AssetParentTrait;